--- lib/middleware/http.d.ts
@@ -1,9 +1,0 @@
-import type { SuperApiToken } from '@prisma/client';
-
-declare module 'http' {
-  interface IncomingMessage {
-    authorizedSpaceId: string;
-    superApiToken: SuperApiToken | null;
-    botUser: User;
-  }
-}
